Refactor into app directory
