+++ mypy.ini
@@ -2,7 +2,7 @@
 
 [mypy]
 python_version = 3.10
-mypy_path = $MYPY_CONFIG_FILE_DIR
+mypy_path = $MYPY_CONFIG_FILE_DIR/app
 exclude = ['venv', '__pycache__']
 plugins = numpy.typing.mypy_plugin
